feat: restructure directories
