+++ src/components/table/config.tsx
