+++ src/components/table/index.tsx
@@ -1,0 +1,3 @@
+export { TableWrapper as Table } from 'components/table/tableWrapper';
+export type { TableWrapperProps as TableProps } from 'components/table/tableWrapper';
+export type { ActionsConfig, ColumnConfig } from 'components/table/config';
