--- src/table/index.tsx
@@ -1,3 +1,0 @@
-export { TableWrapper as Table } from 'table/tableWrapper';
-export type { TableWrapperProps as TableProps } from 'table/tableWrapper';
-export type { ActionsConfig, ColumnConfig } from 'table/config';
